Time: 0 ms (100.00%) | Memory: 19.8 MB (12.43%) - LeetSync

+++ 662-maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -15,7 +15,7 @@ public:
          if (!root)
           return 0;
 
-         queue<pair<TreeNode*, int>> node_queue;
+         queue<pair<TreeNode*, long long>> node_queue;
          node_queue.push(make_pair(root, 0));
 
          int ans = 0;
